requirements.txt
flask==3.0.3
flask-wtf==1.2.1
email-validator==2.2.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-login==0.6.3
python-dotenv==1.0.1
openpyxl==3.1.5
Werkzeug==3.0.4
